refactor: consolidate context_id tests into test_a2a_agent.py
Move TestContextIdPropagation tests from the separate
test_a2a_agent_context_id.py module into the main test_a2a_agent.py
file, per reviewer feedback from @giles17.

- Reuse existing a2a_agent fixture instead of standalone _make_agent()
- Add AgentSession import
- Delete test_a2a_agent_context_id.py
